fix: reactor editor views should word wrap; detail editor views should show line numbers
This also incorporates a small bit of code cleanup to view/detail
