+++ pdl-live-react/src/view/detail/SourceTabContent.tsx
@@ -3,5 +3,5 @@ import type Model from "../timeline/model"
 import Code from "../Code"
 
 export default function SourceTabContent({ block }: { block: Model[number] }) {
-  return <Code block={block.block} />
+  return <Code block={block.block} showLineNumbers />
 }
